feat(Food result)!: Change nutrients to be stored in a dictionary
The nutrient array returned by the API is non-deterministic.
The order of the nutrients in the array can change for each
food. For example, the protein information for one food may be stored
in element 2 while being stored in element 3 for another. This makes it
difficult to select a specific nutrient without manually iterating
over the array.

Using a dictionary ensures that the same line of code will return the
same nutrient type for every food. Each element's key is based off the
"id" value returned for each of the nutrient objects that is then
converted to an enum for easier usage.

BREAKING CHANGE: Nutrients are now accessed using an enum that
represents the nutrient's ID instead of array index